add Volume Info

+++ lib/ClusterParallelCoord/VolumeClustering_Lua.h
@@ -34,8 +34,8 @@ public:
     LUA_SCRIPTCLASS_METHOD_ARG2(LuaTable, GetClusterValue, int, int)
     LUA_SCRIPTCLASS_METHOD_ARG3(int, GetEdgePowers, int, int, int)
     LUA_SCRIPTCLASS_METHOD_ARG2(int, SetSigma, int, float)
-    LUA_SCRIPTCLASS_METHOD_ARG1(int, GetVolumeMin, int)
-    LUA_SCRIPTCLASS_METHOD_ARG1(int, GetVolumeMax, int)
+    LUA_SCRIPTCLASS_METHOD_ARG1(float, GetVolumeMin, int)
+    LUA_SCRIPTCLASS_METHOD_ARG1(float, GetVolumeMax, int)
     
     LUA_SCRIPTCLASS_END();
